requirements.txt
google-generativeai>=0.3.0
gradio>=4.0.0
pandas>=1.5.0
ipywidgets>=8.0.0
